initial game emgine modularity recoding.
